fix: remove old normalization
